Add files via upload
This code compares scipy's newton and fsolve methods with numba's newton method to verify consistent results. The function used is the Hamiltonian for the field line trajectories of magnetic field lines in the non-resonant stellator divertor. The goal is to speed up the time it takes to calculate the next iteration of psi_t.